Move rate functions to separate file
